Connected to Spotify
Set up a basic implicit authentication scheme with spotify.
There are some additional suggestions about using cookies and
using some the optional state param to make it more robust/secure
but I didn't feel it was necessary yet. This approach allows everything
to run client side which is useful at the moment, but has the tradeoff
that the tokens expire after one hour. Probably the regular Authorization Code
approach is more appropriate for a real app.

https://beta.developer.spotify.com/documentation/general/guides/authorization-guide/#implicit-grant-flow

The expected behavior is that to use spotify stuff a user would have
to select 'connect to spotify' which brings up the auth process. Spotify
will be used as the primary song catalog with youtube as a backup. With Spotify
we can specify seek positions with millisecond accuracy which should improve overall
experience. Spotify also allows us to keep the user on the initial page.

The possible scopes list took me a second to find so it's here for future reference
https://beta.developer.spotify.com/documentation/general/guides/scopes/

Seeking turned out to be tricky because the song has to already be playing for it to work
so I put in a sort of event loop to send updates about the current playback state every second
after a song has started playing. It's not perfect, but it was enough to get seek working